feat(l1): reorgs for blocks that are connected to the canonical chain (#577)
**Motivation**

Up until now we can perform reorgs as long as the parent of the newly
added block is canonical. This PR aims to enable blocks that have
canonical ancestors, even if they may not be direct parents.

**Changes**

- `blockchain.add_block`:
- We now stopped checking if the parent is canonical, we rely on the
state transition to check that the parent's state exists in the db,
which is all we need to execute the block.
- The latest canonical block is not updated here now, as it now
symbolizes the head of the chain instead of the latest block that has
arrived.
- `ForkChoiceUpdatedV3`:
  - All ancestors of head are made canonical after the reorg.
- New checks: it is checked that finalized, safe and head blocks are
part of the new canonical chain and in the correct order.
- We now allow for finalized and safe hashes to be zero in the request
(needed for near-merge times when no blocks can be yet dimmed safe or
finalized).
- Most of the fork choice functionality is moved to the blockchain.rs
file and errors are returned when the request should fail and payload
attributes should not be executed.
- Fork choice has now been moved to before payload building, as it needs
to be applied regardless of wether the payload succeeds or not. This is
to be completed by #853 as the parsing needs to be fixed too.
  - Logs are added for easier debugging.
- A basic integration test is added for the payload + fork choice
updated functionality.
- We no longer make imported blocks canonical, even if they come from
the genesis file.

Closes #444